+++ zcash_client_backend/src/proto.rs
@@ -356,8 +356,8 @@ pub enum ProposalDecodingError<DbError> {
     MemoInvalid(memo::Error),
     /// The serialization version returned by the protobuf was not recognized.
     VersionInvalid(u32),
-    /// The proposal did not correctly specify a standard fee rule.
-    FeeRuleNotSpecified,
+    /// The fee rule specified by the proposal is not supported by the wallet.
+    FeeRuleNotSupported(proposal::FeeRule),
     /// The proposal violated balance or structural constraints.
     ProposalInvalid(ProposalError),
     /// An inputs field for the given protocol was present, but contained no input note references.
@@ -409,8 +409,12 @@ impl<E: Display> Display for ProposalDecodingError<E> {
             ProposalDecodingError::VersionInvalid(v) => {
                 write!(f, "Unrecognized proposal version {}", v)
             }
-            ProposalDecodingError::FeeRuleNotSpecified => {
-                write!(f, "Proposal did not specify a known fee rule.")
+            ProposalDecodingError::FeeRuleNotSupported(r) => {
+                write!(
+                    f,
+                    "Fee calculation using the {:?} fee rule is not supported.",
+                    r
+                )
             }
             ProposalDecodingError::ProposalInvalid(err) => write!(f, "{}", err),
             ProposalDecodingError::EmptyShieldedInputs(protocol) => write!(
@@ -600,8 +604,6 @@ impl proposal::Proposal {
         proposal::Proposal {
             proto_version: PROPOSAL_SER_V1,
             fee_rule: match value.fee_rule() {
-                StandardFeeRule::PreZip313 => proposal::FeeRule::PreZip313,
-                StandardFeeRule::Zip313 => proposal::FeeRule::Zip313,
                 StandardFeeRule::Zip317 => proposal::FeeRule::Zip317,
             }
             .into(),
@@ -624,11 +626,9 @@ impl proposal::Proposal {
             PROPOSAL_SER_V1 => {
                 #[allow(deprecated)]
                 let fee_rule = match self.fee_rule() {
-                    proposal::FeeRule::PreZip313 => StandardFeeRule::PreZip313,
-                    proposal::FeeRule::Zip313 => StandardFeeRule::Zip313,
                     proposal::FeeRule::Zip317 => StandardFeeRule::Zip317,
-                    proposal::FeeRule::NotSpecified => {
-                        return Err(ProposalDecodingError::FeeRuleNotSpecified);
+                    other => {
+                        return Err(ProposalDecodingError::FeeRuleNotSupported(other));
                     }
                 };